(tests) fix errors reported 4rm static analysis
issue: #5

+++ lib/net.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 # networking (without sucking)
 #
 # Copyright (c) 2015 GochoMugo <mugo@forfuture.co.ke>
@@ -8,7 +9,10 @@ msu_require "console"
 
 
 # module variables
-DEPS="wget"
+{
+  # shellcheck disable=SC2034
+  DEPS="wget"
+}
 DL_FILE=".msu.download"
 
 
@@ -23,8 +27,9 @@ function track() {
 # check tracked downloads
 function check() {
   if [ -f "${DL_FILE}" ] ; then
-    local names=$(cat "${DL_FILE}" | grep -Eo ".*=" | grep -Eo "[a-Z]*")
-    for name in "${names}"
+    local names
+    names=$(grep -Eo ".*=" "${DL_FILE}" | grep -Eo "[a-Z]*")
+    for name in ${names}
     do
       list "${name}"
     done
@@ -38,39 +43,52 @@ function check() {
 # ${1} - name of download
 # ${2} - url of download (Optional)
 function download() {
-  local name=${1}
-  local url=${2}
-  local data=""
+  local name
+  local url
+  local data
+
+  name=${1}
+  url=${2}
+  data=''
 
   # we require a name
-  if [ ! ${1} ] ; then
+  if [ ! "${1}" ] ; then
     error "we require a name"
     return 1
   fi
 
   # read tracking file, if it exists
-  [ -e "${DL_FILE}" ] && data=$(cat "${DL_FILE}")
+  if [ -e "${DL_FILE}" ]
+  then
+    data=$(cat "${DL_FILE}")
+  fi
 
-  if [ "${url}" ] ; then # new download
+  if [ "${url}" ]
+  then # new download
     track "${name}" "${url}"
-  elif [ "${name}" ]; then # tracked download
+  elif [ "${name}" ]
+  then # tracked download
     url=$(echo "${data}" | grep -E "${name}" | grep -Eo "=.*" | cut -b 2-)
   fi
 
   # download
   if [ "${url}" ] ; then
-    wget -c "${url}" && {
-      echo "${data}" | grep -Ev ${name} > ${DL_FILE}
+    wget -c "${url}"
+    if [ $? ]
+    then
+      echo "${data}" | grep -Ev "${name}" > "${DL_FILE}"
       tick "${name}"
-    } || {
+    else
       cross "${name} (could not download)"
-    }
+    fi
   else
-    echo "${data}" | grep -Eo "=.*" | cut -b 2- | xargs -I{} wget -c {} && {
+    echo "${data}" | grep -Eo "=.*" | cut -b 2- | xargs -I{} wget -c {}
+    if [ $? ]
+    then
       rm -rf "${DL_FILE}"
       success "done"
-    } || {
+    else
       error "did not complete successfully"
-    }
+    fi
   fi
 }
